+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### 1.2.0
 
-- new configuration option `implicit_validation` +
+- new configuration option `implicit_validation`, new method `Essence::implicitValidation` +
 - new validation methods `Essence::validateAll`, `Essence::validate` (alias of `Essence::go`) +
 - new configuration methods `Essence::addLink` and `Essence::addMatcher` +
 - wrapper `expect` now works with other data types (not only `Essence` instances) +